Extract terminal file-upload logic into terminal-uploads.ts
use-terminal.ts is a 916-line complexity hotspot. The self-contained
drag/drop + paste upload orchestration (validation, upload loop, toast
feedback) had no dependency on the terminal connection lifecycle, so it
moves cleanly into a new terminal-uploads.ts module — continuing the
established extraction pattern (terminal-socket.ts #659, terminal-surface.ts
#664). Pure code move: the hook's uploadFiles callback now delegates to
uploadTerminalFiles(). No behavior change.

Hook shrinks 916 -> 867 lines and sheds the sonner/media-upload imports.
